ddsBackend/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.4.2
autopep8==1.5.5
colorama==0.4.4
distlib==0.3.1
Django==3.2
django-filter==2.4.0
filelock==3.0.12
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.3
Pillow==8.2.0
pycodestyle==2.6.0
pylint==2.6.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.4
virtualenvwrapper-win==1.2.6
wrapt==1.12.1
